Updates the about me widget image + adds 250x250 emojis
