+++ src/components/widgets/aboutMe/aboutMe.scss
@@ -1,0 +1,7 @@
+.aboutMeImage {
+    margin: 0 auto;
+}
+
+.aboutMeImageDivider {
+    margin-top: 0;
+}
